# binctl_server/openapi.yaml
openapi: 3.0.3
info:
  title: binctl
  version: "1.0.0"
servers:
  - url: /

security:
  - bearerAuth: []

paths:
  /v1/nodes:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_nodes_list
      tags: [nodes]
      summary: List all nodes
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePage"
    post:
      x-openapi-router-controller: binctl_server.routes
      operationId: post_node_create
      tags: [nodes]
      summary: Create a new node
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"

  /v1/nodes/{node_id}:
    parameters:
      - in: path
        name: node_id
        required: true
        schema:
          type: string
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_node_detail
      tags: [nodes]
      summary: Get a node by ID
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Node not found
    patch:
      x-openapi-router-controller: binctl_server.routes
      operationId: patch_node_update
      tags: [nodes]
      summary: Update an existing node
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeUpdate"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Node not found
    put:
      x-openapi-router-controller: binctl_server.routes
      operationId: put_node_upsert
      tags: [nodes]
      summary: Create or fully replace a node with a client-supplied ID
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeReplace"
      responses:
        "200":
          description: Node replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    delete:
      x-openapi-router-controller: binctl_server.routes
      operationId: delete_node_endpoint
      tags: [nodes]
      summary: Delete a node
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Node deleted
          content:
            application/json:
              schema:
                type: object
                required: 
                  - deleted
                properties:
                  deleted:
                    type: object
                    required:
                      - total
                      - edges
                      - tags
                      - nodes
                    properties:
                      total:
                        type: integer
                      edges:
                        type: integer
                      tags:
                        type: integer
                      nodes:
                        type: integer
        "404":
          description: Node not found


  /v1/tags:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_tags_list
      tags: [tags]
      summary: List all tags
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagPage"
    post:
      x-openapi-router-controller: binctl_server.routes
      operationId: post_tag_create
      tags: [tags]
      summary: Create a tag
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreate"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"

  /v1/tags/{tag_name}:
    parameters:
      - in: path
        name: tag_name
        required: true
        schema:
          type: string
          pattern: "^[a-z-]+$"
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_tag_detail
      tags: [tags]
      summary: Get tag by name
      responses:
        "200":
          description: Tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "404":
          description: Tag not found
    patch:
      x-openapi-router-controller: binctl_server.routes
      operationId: patch_tag_update
      tags: [tags]
      summary: Update a tag
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "404":
          description: Tag not found
    delete:
      x-openapi-router-controller: binctl_server.routes
      operationId: delete_tag_endpoint
      tags: [tags]
      summary: Delete a tag
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Tag deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                properties:
                  deleted:
                    type: object
                    required:
                      - total
                      - associations
                    properties:
                      total:
                        type: integer
                      associations:
                        type: integer
        "404":
          description: Tag not found

  /v1/lock:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_lock
      tags: [config]
      summary: Get database-wide idempotency mode
      responses:
        "200":
          description: Current mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdempotencyLock"
    put:
      x-openapi-router-controller: binctl_server.routes
      operationId: enable_lock
      tags: [config]
      summary: Enable database-wide idempotency mode
      responses:
        "200":
          description: Mode enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdempotencyLock"
    delete:
      x-openapi-router-controller: binctl_server.routes
      operationId: disable_lock
      tags: [config]
      summary: Disable database-wide idempotency mode
      responses:
        "200":
          description: Mode disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdempotencyLock"

  /v1/snapshot:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_snapshot
      tags: [nodes]
      summary: Download a snapshot of the current graph
      responses:
        "200":
          description: Point-in-time graph snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphSnapshot"

  /v1/events:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_events
      tags: [events]
      summary: Replay inventory events and continue streaming live updates
      parameters:
        - in: header
          name: Last-Event-ID
          required: true
          schema:
            type: string
            pattern: "^[0-9]+$"
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Invalid event cursor

  /v1/config:
    get:
      x-openapi-router-controller: binctl_server.routes
      operationId: get_config
      tags: [config]
      summary: Get server configuration
      responses:
        "200":
          description: Server configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerConfig"

  /v1/auth/login:
    post:
      x-openapi-router-controller: binctl_server.auth_operations
      operationId: login
      tags: [auth]
      summary: Obtain a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /v1/auth/logout:
    post:
      x-openapi-router-controller: binctl_server.auth_operations
      operationId: logout
      tags: [auth]
      summary: Revoke the current bearer token
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out successfully

components:
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: binctl_server.auth_operations.lookup_token

  schemas:

    IdempotencyLock:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean

    GraphSnapshot:
      type: object
      required: [nodes, event_cursor]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        event_cursor:
          type: integer
          minimum: 0

    InventoryEventData:
      type: object
      required: [occurred_at, resource_type, operation]
      properties:
        occurred_at:
          type: string
          format: date-time
        resource_type:
          type: string
          enum: [node, tag]
        operation:
          type: string
          enum: [created, updated, deleted]
        resource_id:
          type: string
        previous_name:
          type: string
        resource:
          type: object
        deleted:
          type: object

    ResetRequiredEventData:
      type: object
      required: [reason, current_cursor]
      properties:
        reason:
          type: string
          enum: [event-history-expired]
        current_cursor:
          type: integer
          minimum: 0

    NodePage:
      type: object
      required: [total, limit, offset, items]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Node"

    TagPage:
      type: object
      required: [total, limit, offset, items]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tag"

    NodeCreate:
      type: object
      required: [label]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        is_container:
          type: boolean
          default: false
        parent_id:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          default: []

    NodeReplace:
      type: object
      required: [label]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        is_container:
          type: boolean
          default: false
        parent_id:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 255
          default: []

    NodeUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        is_container:
          type: boolean
        parent_id:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 255

    Node:
      type: object
      required: [id, label, description, is_container, created_at, updated_at, parent_id, tags]
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
        is_container:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        parent_id:
          type: string
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        tags:
          type: array
          items:
            type: string

    TagCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-z-]+$"
    TagUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-z-]+$"

    Tag:
      type: object
      required: [name, created_at, updated_at]
      properties:
        name:
          type: string
          pattern: "^[a-z-]+$"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string

    ServerConfig:
      type: object
      required: [orphan_location]
      properties:
        orphan_location:
          type: string
          nullable: true